+++ Changey/Services/IChangeLogReleaser.cs
@@ -4,6 +4,6 @@ namespace Changey.Services
 {
 	internal interface IChangeLogReleaser
 	{
-		Task<bool> Release(string path, string? date, string version);
+		Task<bool> Release(string path, string? date, string version, bool forceRelease);
 	}
 }